Finished importing data in postgres
