+++ db/Postgres/postgSchema.sql
@@ -2,6 +2,8 @@ DROP DATABASE IF EXISTS aboutInfo;
 CREATE DATABASE aboutInfo;
 
 DROP TABLE IF EXISTS hosts CASCADE;
+DROP TABLE IF EXISTS listing CASCADE;
+DROP TABLE IF EXISTS ratings CASCADE;
 
 CREATE TABLE hosts (
   id SERIAL PRIMARY KEY,
@@ -19,29 +21,54 @@ CREATE TABLE hosts (
   response_time integer DEFAULT 0,
   languages varchar(255) NOT NULL,
   email varchar(255) NOT NULL
+  
+);
+
+
+CREATE TABLE listing (
+  id SERIAL PRIMARY KEY,
+  features varchar(500),
+  things_to_do varchar(500),
+  lat_location decimal(3),
+  lon_location decimal(3),
+  host_guidebook varchar(255)
+
+);
+
+
+CREATE TABLE ratings (
+  id SERIAL PRIMARY KEY,
+  rating integer NOT NULL,
+  host_id integer REFERENCES hosts(id),
+  listing_id integer REFERENCES listing(id)
+
 );
 
 
 
--- need to change url names + get rid of headers in csv files
+
 
 copy hosts (list_name, first_name, last_name, city, 
 	state, country, joined_in_date, references_count, verified, 
 	description, response_rate, response_time, languages, email) 
-from '/Users/gauravgulati/desktop/SDC/about-service/db/Postgres/csvFiles/firstFourthOfHostTable.csv' DELIMITERS ',' CSV;
+from '/Users/gauravgulati/desktop/SDC/about-service/db/Postgres/csvFiles/host/firstFourthOfHostTable.csv' DELIMITERS ',' CSV;
 
 copy hosts (list_name, first_name, last_name, city, 
 	state, country, joined_in_date, references_count, verified, 
 	description, response_rate, response_time, languages, email) 
-from '/Users/gauravgulati/desktop/SDC/about-service/db/Postgres/csvFiles/secondFourthOfHostTable.csv' DELIMITERS ',' CSV;
+from '/Users/gauravgulati/desktop/SDC/about-service/db/Postgres/csvFiles/host/secondFourthOfHostTable.csv' DELIMITERS ',' CSV;
 
 copy hosts (list_name, first_name, last_name, city, 
 	state, country, joined_in_date, references_count, verified, 
 	description, response_rate, response_time, languages, email) 
-from '/Users/gauravgulati/desktop/SDC/about-service/db/Postgres/csvFiles/thirdFourthOfHostTable.csv' DELIMITERS ',' CSV;
+from '/Users/gauravgulati/desktop/SDC/about-service/db/Postgres/csvFiles/host/thirdFourthOfHostTable.csv' DELIMITERS ',' CSV;
 
 copy hosts (list_name, first_name, last_name, city, 
 	state, country, joined_in_date, references_count, verified, 
 	description, response_rate, response_time, languages, email) 
-from '/Users/gauravgulati/desktop/SDC/about-service/db/Postgres/csvFiles/fourthFourthOfHostTable.csv' DELIMITERS ',' CSV;
+from '/Users/gauravgulati/desktop/SDC/about-service/db/Postgres/csvFiles/host/fourthFourthOfHostTable.csv' DELIMITERS ',' CSV;
 
+copy listing (features, things_to_do, lat_location, lon_location) from '/Users/gauravgulati/desktop/SDC/about-service/db/Postgres/csvFiles/listings/first_listingsTableData.csv' DELIMITERS ',' CSV;
+copy listing (features, things_to_do, lat_location, lon_location) from '/Users/gauravgulati/desktop/SDC/about-service/db/Postgres/csvFiles/listings/second_listingsTableData.csv' DELIMITERS ',' CSV;
+
+copy ratings(rating, host_id, listing_id) from '/Users/gauravgulati/desktop/SDC/about-service/db/Postgres/csvFiles/ratings/ratingsTable.csv' DELIMITERS ',' CSV;
