+++ package.json
@@ -20,6 +20,7 @@
     "body-parser": "^1.18.3",
     "csv-write-stream": "^2.0.0",
     "express": "^4.16.2",
+    "faker": "^4.1.0",
     "google-map-react": "^1.0.5",
     "jquery": "^3.3.1",
     "morgan": "^1.9.0",
